+++ Backend/estate-emporium/Models/db/PropertySale.cs
